Add files via upload
The following project is part of a feasibility analysis of a taxi service business that I was looking to start. In the worksheet, you will see different models that highlight startup costs, monthly costs, and revenue projections, which are based on how many short and long term rides that are expected each month. In the additional worksheets, I also analyzed how much  potential revenue increase there could be by utilizing a uber style pricing strategy verse a fixed pricing strategy, and what the potential cost difference would be if we used an electric car, and changed different parts of how the business was ran. 

With a projected ceiling of about 100 rides per month, one can see that there is not enough revenue to successfully run the business. I then presented my findings to my professor and team members who were also part of the project.